requirements.txt
flask==2.0.1
werkzeug==2.0.1
kafka-python==2.0.2
python-dotenv==0.19.0
pandas==1.3.3
numpy==1.21.2 